+++ CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
 - add command-line search flags --limit and --offset
 - tweak icon borders
 - fix command-line search ignoring --type flag
+- fix command-line search output for chats, contacts, and tables
 - fix error on using system colours with transparency
 - fix errors on Ubuntu 22 (issue #111)